pyproject.toml
[tool.poetry]
name = "example-app"
version = "0.1.0"
description = ""
authors = ["Matt Wright <mdw1980@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
litestar = "^2.5.1"
pycrdt = "^0.8.6"
pycrdt-websocket = "^0.12.6"
uvicorn = "^0.27.0"
websockets = "^12.0"
redis = "^5.0.1"
structlog = "^24.1.0"
jinja2 = "^3.1.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-playwright = "^0.4.3"
testcontainers-redis = "^0.0.1rc1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
